added some features in admin side Ui

+++ lib/features/orders/data/models/order_details_model.dart
@@ -50,8 +50,8 @@ class OrderDetailsModel {
       paymentMethod: (json['payment']?['method'] ?? 'COD').toString(),
       paymentStatus: (json['payment']?['status'] ?? 'pending').toString(),
       subtotal: subtotal,
-      tax: tax,
-      shippingFee: shipping,
+      tax: tax.toDouble(),
+      shippingFee: shipping.toDouble(),
       total: total,
     );
   }
